# internal/templates/html/static/openapi.yaml
openapi: 3.0.3
info:
  title: Notebird API
  version: 0.1.0
  description: Local HTTP API for Notebird, a personal wiki built from Chirps.
servers:
  - url: http://127.0.0.1:7331
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        '200':
          description: Server is healthy
          content:
            text/plain:
              schema:
                type: string
                example: ok
  /:
    get:
      summary: Render the main Notebird UI
      responses:
        '200':
          description: HTML application shell
          content:
            text/html:
              schema:
                type: string
  /chirps:
    get:
      summary: Render the Notebird UI with a filtered feed
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/TagFilter'
        - $ref: '#/components/parameters/FeedMode'
      responses:
        '200':
          description: HTML application shell
          content:
            text/html:
              schema:
                type: string
    post:
      summary: Create a Chirp
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [text]
              properties:
                title:
                  type: string
                  description: Optional title. Defaults to frontmatter title or first line of text.
                text:
                  type: string
                  description: Markdown body. YAML-style frontmatter is parsed and pruned before storage.
                tags:
                  type: string
                  description: Optional space- or comma-separated tags from the tag input.
      responses:
        '200':
          description: Feed HTML fragment after creation
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Invalid form data
  /chirps/{id}:
    get:
      summary: Render the Notebird UI with a Chirp selected
      parameters:
        - $ref: '#/components/parameters/ChirpID'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/TagFilter'
        - $ref: '#/components/parameters/FeedMode'
      responses:
        '200':
          description: HTML application shell
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Chirp not found
    put:
      summary: Update a Chirp
      parameters:
        - $ref: '#/components/parameters/ChirpID'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [text]
              properties:
                title:
                  type: string
                text:
                  type: string
                  description: Markdown body. YAML-style frontmatter is parsed and pruned before storage.
                tags:
                  type: string
                  description: Optional space- or comma-separated tags from the tag input.
      responses:
        '200':
          description: Create composer HTML fragment after update
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Invalid form data
    delete:
      summary: Delete a Chirp
      parameters:
        - $ref: '#/components/parameters/ChirpID'
      responses:
        '200':
          description: Empty detail HTML fragment
          content:
            text/html:
              schema:
                type: string
  /chirps/{id}/edit:
    get:
      summary: Render the Notebird UI with a Chirp loaded in the composer
      parameters:
        - $ref: '#/components/parameters/ChirpID'
      responses:
        '200':
          description: HTML application shell
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Chirp not found
  /x/chirps:
    get:
      summary: Render feed partial
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/TagFilter'
        - $ref: '#/components/parameters/FeedMode'
        - $ref: '#/components/parameters/JSONPartial'
      responses:
        '200':
          description: Feed partial as HTML or JSON envelope
          content:
            text/html:
              schema:
                type: string
            application/json:
              schema:
                $ref: '#/components/schemas/PartialResponse'
  /x/chirps/{id}:
    get:
      summary: Render Chirp detail partial
      parameters:
        - $ref: '#/components/parameters/ChirpID'
        - $ref: '#/components/parameters/JSONPartial'
      responses:
        '200':
          description: Detail partial as HTML or JSON envelope
          content:
            text/html:
              schema:
                type: string
            application/json:
              schema:
                $ref: '#/components/schemas/PartialResponse'
        '404':
          description: Chirp not found
  /x/chirps/{id}/edit:
    get:
      summary: Render Chirp edit composer partial
      parameters:
        - $ref: '#/components/parameters/ChirpID'
        - $ref: '#/components/parameters/JSONPartial'
      responses:
        '200':
          description: Edit composer partial as HTML or JSON envelope
          content:
            text/html:
              schema:
                type: string
            application/json:
              schema:
                $ref: '#/components/schemas/PartialResponse'
        '404':
          description: Chirp not found
  /x/composer:
    get:
      summary: Render create composer partial
      parameters:
        - $ref: '#/components/parameters/JSONPartial'
      responses:
        '200':
          description: Composer partial as HTML or JSON envelope
          content:
            text/html:
              schema:
                type: string
            application/json:
              schema:
                $ref: '#/components/schemas/PartialResponse'
  /x/nav:
    get:
      summary: Render sidebar navigation partial
      parameters:
        - $ref: '#/components/parameters/JSONPartial'
      responses:
        '200':
          description: Sidebar partial as HTML or JSON envelope
          content:
            text/html:
              schema:
                type: string
            application/json:
              schema:
                $ref: '#/components/schemas/PartialResponse'
  /preview:
    post:
      summary: Render Markdown preview
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  description: Markdown body to preview.
      responses:
        '200':
          description: Rendered Markdown HTML fragment
          content:
            text/html:
              schema:
                type: string
  /suggest:
    get:
      summary: Suggest Chirp titles or tags for editor completion
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [chirp, tag]
            default: chirp
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Completion suggestions
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Suggestion'
  /debug/config:
    get:
      summary: Show runtime config
      responses:
        '200':
          description: Runtime config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
  /debug/vars:
    get:
      summary: expvar debug variables
      responses:
        '200':
          description: expvar JSON
          content:
            application/json:
              schema:
                type: object
  /docs:
    get:
      summary: Swagger UI API documentation
      responses:
        '200':
          description: API documentation page
          content:
            text/html:
              schema:
                type: string
  /openapi.yaml:
    get:
      summary: OpenAPI v3 specification
      responses:
        '200':
          description: OpenAPI document
          content:
            application/yaml:
              schema:
                type: string
components:
  parameters:
    ChirpID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Chirp ULID
    SearchQuery:
      name: q
      in: query
      schema:
        type: string
      description: Full-text search query.
    TagFilter:
      name: tag
      in: query
      schema:
        type: string
      description: Filter feed by tag.
    FeedMode:
      name: mode
      in: query
      schema:
        type: string
        enum: [timeline, wanted]
      description: Feed mode.
    JSONPartial:
      name: json
      in: query
      schema:
        type: string
        enum: ['1', 'true', 'yes']
      description: When present and truthy, /x/ partial routes return a JSON envelope with rendered HTML and backing data.
  schemas:
    PartialResponse:
      type: object
      properties:
        template:
          type: string
          description: Template name rendered for the partial.
        html:
          type: string
          description: Rendered HTML for the partial.
        data:
          type: object
          additionalProperties: true
          description: Data object passed to the template.
    Suggestion:
      type: object
      properties:
        label:
          type: string
        value:
          type: string
        detail:
          type: string
    Config:
      type: object
      properties:
        data_dir:
          type: string
        host:
          type: string
        port:
          type: integer
    Chirp:
      type: object
      properties:
        id:
          type: string
          description: ULID canonical identity.
        title:
          type: string
        text:
          type: string
        type:
          type: string
          example: text/markdown
        tags:
          type: array
          items:
            type: string
        fields:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time